Create Credx_FinancialAnalysis.ipynb
Assessing the financial benefit of your project
You need to assess and explain the potential financial benefit of your project to the bank's management. From a P&L perspective, identify the metrics you are trying to optimise, explain (in simple terms) how the analysis and the model work, and share the results of the model. Finally, assess the financial benefit of the model and report the following:

The implications of using the model for auto-approval or rejection, i.e., how many applicants on an average would the model automatically approve or reject

The potential credit loss avoided with the help of the model

Assumptions based on which the model was built